requirements.txt
pandas==2.2.2
requests==2.31.0
pillow==10.3.0